Update Checkout.php
The current 'trim()' function is throwing the following error in the 'var/log/debug.log':

trim(): Passing null to parameter #1 ($string) of type string is deprecated in /var/www/magento/html/app/code/Tco/Checkout/Helper/Checkout.php on line 188

I think it's related to the php version 8.1 that I am using on Magento 2.4.6.

Here is the line of code in question:

$trimmedValue = trim($value);

This error raised in 'debug.log' prevents the checkout process from transferring from the Magento website to 2checkout hosted page, thus preventing the completion of a transaction via hosted payments method

When changed to 'trim($value ?? '');' the bug is fixed and hosted payment checkout process is successfully passed from Magento website to 2checkout hosted page.

this modification uses the null operator (??), which will use an empty string as a fallback if $value is null. This way, you're ensuring that the trim() function receives a string as its parameter, avoiding the deprecated functionality error.